+++ psbuilder-script.py
@@ -2,10 +2,11 @@
 import sys
 from pypsbuilder import psbuilder
 
+
 def main():
     print('Running psbuilder...')
     sys.exit(psbuilder.main())
 
+
 if __name__=='__main__':
     main()
-