+++ psiso-script.py
@@ -3,6 +3,7 @@ import sys
 import argparse
 from pypsbuilder.psexplorer import PTPS
 
+
 def main():
     parser = argparse.ArgumentParser(description='Draw isopleth diagrams')
     parser.add_argument('project', type=str,
@@ -17,12 +18,22 @@ def main():
                         default=None, help='contour step')
     parser.add_argument('--ncont', type=int,
                         default=10, help='number of contours')
+    parser.add_argument('--colors', type=str,
+                        default=None, help='color for all levels')
+    parser.add_argument('--cmap', type=str,
+                        default=None, help='name of the colormap')
     parser.add_argument('--smooth', type=float,
                         default=0, help='smoothness of the approximation')
+    parser.add_argument('--clabel', nargs='+',
+                        default=[], help='label contours in field defined by set of phases')
     args = parser.parse_args()
     print('Running psiso...')
     ps = PTPS(args.project)
-    sys.exit(ps.isopleths(args.phase, args.expr, filled=args.filled, smooth=args.smooth, step=args.step, N=args.ncont))
+    sys.exit(ps.isopleths(args.phase, args.expr, filled=args.filled,
+                          smooth=args.smooth, step=args.step,
+                          N=args.ncont, clabel=args.clabel,
+                          colors=args.colors, cmap=args.cmap))
+
 
 if __name__ == "__main__":
     main()